Separate supabase fake classes from main wrapper class
