+++ lib/libraries/supabase/testing/fake_supabase_wrapper.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:io';
 import 'package:construculator/libraries/supabase/interfaces/supabase_wrapper.dart';
+import 'package:construculator/libraries/supabase/testing/fake_supabase_classes.dart';
 import 'package:supabase_flutter/supabase_flutter.dart' as supabase;
 
 /// Fake implementation of ISupabaseWrapper for testing
@@ -436,120 +437,3 @@ class FakeSupabaseWrapper implements SupabaseWrapper {
     simulateAuthStreamError(errorMessage);
   }
 }
-
-// Fake exception classes that mimic Supabase exceptions
-class FakeAuthException implements supabase.AuthException {
-  @override
-  final String message;
-  
-  @override
-  final String? statusCode;
-
-  @override
-  final String? code;
-
-  FakeAuthException(this.message, {this.statusCode, this.code});
-
-  @override
-  String toString() => message;
-
-  @override
-  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
-}
-
-class FakePostgrestException implements supabase.PostgrestException {
-  @override
-  final String message;
-  
-  @override
-  final String? code;
-  
-  @override
-  final String? details;
-  
-  @override
-  final String? hint;
-
-  FakePostgrestException(this.message, {this.code, this.details, this.hint});
-
-  @override
-  String toString() => message;
-
-  @override
-  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
-}
-
-class FakeSession implements supabase.Session {
-  @override
-  final supabase.User user;
-  
-  @override
-  final String accessToken;
-  
-  @override
-  final String refreshToken;
-
-  FakeSession({
-    required this.user,
-    this.accessToken = 'fake-access-token',
-    this.refreshToken = 'fake-refresh-token',
-  });
-
-  @override
-  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
-}
-
-// Fake implementations of Supabase classes
-class FakeUser implements supabase.User {
-  @override
-  final String id;
-  
-  @override
-  final String? email;
-  
-  @override
-  final String createdAt;
-  
-  @override
-  final Map<String, dynamic> appMetadata;
-  
-  @override
-  final Map<String, dynamic>? userMetadata;
-
-  FakeUser({
-    required this.id,
-    this.email,
-    required this.createdAt,
-    this.appMetadata = const {},
-    this.userMetadata,
-  });
-
-  @override
-  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
-}
-
-class FakeAuthResponse implements supabase.AuthResponse {
-  @override
-  final supabase.User? user;
-  
-  @override
-  final supabase.Session? session;
-
-  FakeAuthResponse({this.user, this.session});
-
-  @override
-  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
-}
-
-class FakeAuthState implements supabase.AuthState {
-  @override
-  final supabase.AuthChangeEvent event;
-  
-  @override
-  final supabase.Session? session;
-
-  FakeAuthState({required this.event, this.session});
-
-  @override
-  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
-}
